# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BillListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BillResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "BillListResponse",
        "type": "object"
      },
...
      "BillingListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BillingListItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "stats": {
            "$ref": "#/components/schemas/BillingStatsResponse"
          },
          "user_pix_incomplete": {
            "title": "User Pix Incomplete",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "user_pix_incomplete",
          "stats"
        ],
        "title": "BillingListResponse",
        "type": "object"
      },
...
      "ContactListRequest": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContactInput"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ContactListRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/billings": {
      "get": {
        "operationId": "list_billings_api_v1_billings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Billings",
        "tags": [
          "billings"
        ]
      },
      "post": {
        "operationId": "create_billing_api_v1_billings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Billing",
        "tags": [
          "billings"
        ]
      }
    },
    "/api/v1/billings/{billing_uuid}": {
      "delete": {
        "operationId": "delete_billing_api_v1_billings__billing_uuid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Billing",
        "tags": [
          "billings"
        ]
      },
      "get": {
        "operationId": "get_billing_api_v1_billings__billing_uuid__get",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Billing",
        "tags": [
          "billings"
        ]
      },
      "patch": {
        "operationId": "update_billing_api_v1_billings__billing_uuid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Billing",
        "tags": [
          "billings"
        ]
      }
    },
...
    "/api/v1/billings/{billing_uuid}/attachments/{attachment_uuid}": {
      "delete": {
        "operationId": "delete_attachment_api_v1_billings__billing_uuid__attachments__attachment_uuid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "attachment_uuid",
            "required": true,
            "schema": {
              "title": "Attachment Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Attachment",
        "tags": [
          "billings"
        ]
      },
      "get": {
        "operationId": "download_attachment_api_v1_billings__billing_uuid__attachments__attachment_uuid__get",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "attachment_uuid",
            "required": true,
            "schema": {
              "title": "Attachment Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Attachment content"
          },
          "302": {
            "description": "Temporary storage redirect",
            "headers": {
              "Location": {
                "description": "Signed attachment URL",
                "schema": {
                  "format": "uri",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Attachment",
        "tags": [
          "billings"
        ]
      }
    },
    "/api/v1/billings/{billing_uuid}/bills": {
      "get": {
        "operationId": "list_bills_api_v1_billings__billing_uuid__bills_get",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Bills",
        "tags": [
          "bills"
        ]
      },
      "post": {
        "operationId": "create_bill_api_v1_billings__billing_uuid__bills_post",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillCreateRequest"
              }
            },
            "multipart/form-data": {
              "encoding": {
                "payload": {
                  "contentType": "application/json"
                }
              },
              "schema": {
                "$ref": "#/components/schemas/Body_create_bill_api_v1_billings__billing_uuid__bills_post",
                "properties": {
                  "payload": {
                    "$ref": "#/components/schemas/BillCreateRequest"
                  },
                  "receipt_files": {
                    "items": {
                      "format": "binary",
                      "type": "string"
                    },
                    "type": "array"
                  }
                },
                "required": [
                  "payload"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Create Bill",
        "tags": [
          "bills"
        ]
      }
    },
...
    "/api/v1/billings/{billing_uuid}/recipients": {
      "put": {
        "operationId": "replace_recipients_api_v1_billings__billing_uuid__recipients_put",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Recipients",
        "tags": [
          "billings"
        ]
      }
    },
...
    "/api/v1/billings/{billing_uuid}/transfer": {
      "post": {
        "operationId": "transfer_billing_api_v1_billings__billing_uuid__transfer_post",
        "parameters": [
          {
            "in": "path",
            "name": "billing_uuid",
            "required": true,
            "schema": {
              "title": "Billing Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/rentivo__api__schemas__billings__BillingTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transfer Billing",
        "tags": [
          "billings"
        ]
      }
    },
...
  }
}